Trace synchronous Windows ARM startup
